+++ test/scripts/utils/types.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright © 2022 Lisk Foundation
+ *
+ * See the LICENSE file at the top-level directory of this distribution
+ * for licensing information.
+ *
+ * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
+ * no part of this software, including this file, may be copied, modified,
+ * propagated, or distributed except according to the terms contained in the
+ * LICENSE file.
+ *
+ * Removal or modification of this copyright notice is prohibited.
+ */
+
+export interface PassphraseAndKeys {
+	passphrase: string;
+	privateKey?: Buffer;
+	publicKey: Buffer;
+	address: string;
+}
+
+export interface Account {
+	passphrase: string;
+	privateKey: Buffer;
+	publicKey: Buffer;
+	address: string;
+	blsKey: Buffer;
+	generatorKey: Buffer;
+	proofOfPossession: Buffer;
+}
